requirements.txt
# Supabase 客戶端
supabase==2.0.0

# OpenAI API
openai==1.3.0

# 環境變數管理
python-dotenv==1.0.0

# 資料處理
pandas==2.1.3

# 進度條
tqdm==4.66.1

# 重試機制
tenacity==8.2.3

# 資料驗證
pydantic==2.5.0

# 日誌
loguru==0.7.2

# 測試
pytest==7.4.3
pytest-asyncio==0.21.1

# 型別檢查
mypy==1.7.1

